Fix entity error
